Fixing the way events in the past are updated
When placing an event in the past, lastprio and buckets were updated
before executing a possible resize() operation. resize() in its turn
calls calqueue_deq(), which has a wrong behaviour if lastprio and
buckets were not correctly updated.
This commit changes the order according to which the operations are
carried out.

Signed-off-by: Alessandro Pellegrini <alessandro@pellegrini.tk>