Added executable flag to reverse window
A call to mprotect has been added within the revwin creation in order to
give the executable section of the revwin grant to be executed.

Still, to be fixed:
Note that mproptect wants a page-aligned address, though the window is handled
by the SLAB allocator which might retur a non-aligned address for it.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>