Fixed a mprotect issue in the slab allocator
The current implementation of the slab allocator was not desinged to
properly give executable permissions to the newly allocated slab,
i.e. the reverse window.

mprotect was removed from reverse.c and adde to the slab allocator.

Further a REVERSE flag has been added to the core.h to enable reversible
execution.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>