Added a clock counter to statistics
The function gettimeofday was not safe to be used in transactional
contexts, therefore a new timer has been introduced in order to keep
track of clock cycles, instead.

Accordingly to the new clock timer, statistics gathering in the main
loop on core.c has been modified. Specifically, HTM and STM executions
have a set of possible execution branches which was not be taken into
account, during the time measurement.

Signed-off-by: Davide Cingolani <cingolani@dis.uniroma1.it>