Created a HASH model as variant of PHOLD
HASH model is a variant of PHOLD which simulates memory accesses,
both in read and write mode. The latancy of each event can be adjusted
with a variance value.

The model is intended to verify reverse module to work properly by
computing the hash of the whole LP's state at each event.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>