Add print to debug
Add a print primitive for a mode comprehensive output with a more
compact syntax.

Comment out part of the ONGVT section. The section did a force rollback
to a commit horizion state, hence realign the system to the bound event;
however, the LP state OnGVT function will be passed with is not the
platform one, therefore it is safe to pass it by the pointer in the LP
structure relying on the fact that rollback will do everything right.

Davide Cingolani <cingolani@diag@uniroma1.it>