Add checkpoint recalculation
New model to recalculate checkpoint period is added. The function will
be periodically called to recompute the best period based on the Lin's
model. In order to do this, some statistics have to be online.

At the same time, statistics subsystem has been updated and realigend to
the current state of the code.